rename files to snake case and change branch fetch to 25 from 50
