+++ src/sentry/codecov/endpoints/Repositories/repositories.py
@@ -10,8 +10,8 @@ from sentry.apidocs.constants import RESPONSE_BAD_REQUEST, RESPONSE_FORBIDDEN, R
 from sentry.apidocs.parameters import GlobalParams, PreventParams
 from sentry.codecov.base import CodecovEndpoint
 from sentry.codecov.client import CodecovApiClient
-from sentry.codecov.endpoints.Repositories.query import query
-from sentry.codecov.endpoints.Repositories.serializers import RepositoriesSerializer
+from sentry.codecov.endpoints.repositories.query import query
+from sentry.codecov.endpoints.repositories.serializers import RepositoriesSerializer
 from sentry.codecov.enums import NavigationParameter, OrderingDirection
 from sentry.integrations.services.integration.model import RpcIntegration
