+++ tests/sentry/codecov/endpoints/test_test_result_aggregates.py
@@ -32,7 +32,7 @@ class TestResultsAggregatesEndpointTest(APITestCase):
         )
 
     @patch(
-        "sentry.codecov.endpoints.TestResultsAggregates.test_results_aggregates.CodecovApiClient"
+        "sentry.codecov.endpoints.test_results_aggregates.test_results_aggregates.CodecovApiClient"
     )
     def test_get_returns_mock_response(self, mock_codecov_client_class):
         mock_graphql_response = {
@@ -94,7 +94,7 @@ class TestResultsAggregatesEndpointTest(APITestCase):
         assert response.data["flakeRatePercentChange"] == 0.1
 
     @patch(
-        "sentry.codecov.endpoints.TestResultsAggregates.test_results_aggregates.CodecovApiClient"
+        "sentry.codecov.endpoints.test_results_aggregates.test_results_aggregates.CodecovApiClient"
     )
     def test_get_with_interval_query_param(self, mock_codecov_client_class):
         mock_graphql_response = {
