+++ tests/sentry/codecov/endpoints/test_test_results.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock, patch
 
 from django.urls import reverse
 
-from sentry.codecov.endpoints.TestResults.serializers import (
+from sentry.codecov.endpoints.test_results.serializers import (
     TestResultNodeSerializer as NodeSerializer,
 )
 from sentry.testutils.cases import APITestCase
@@ -88,7 +88,7 @@ mock_graphql_response_populated = {
 
 
 class TestResultsEndpointTest(APITestCase):
-    endpoint = "sentry-api-0-test-results"
+    endpoint_name = "sentry-api-0-test-results"
 
     def setUp(self):
         super().setUp()
@@ -104,7 +104,7 @@ class TestResultsEndpointTest(APITestCase):
     def reverse_url(self, owner="testowner", repository="testrepo"):
         """Custom reverse URL method to handle required URL parameters"""
         return reverse(
-            self.endpoint,
+            self.endpoint_name,
             kwargs={
                 "organization_id_or_slug": self.organization.slug,
                 "owner": self.integration.id,
@@ -112,7 +112,7 @@ class TestResultsEndpointTest(APITestCase):
             },
         )
 
-    @patch("sentry.codecov.endpoints.TestResults.test_results.CodecovApiClient")
+    @patch("sentry.codecov.endpoints.test_results.test_results.CodecovApiClient")
     def test_get_returns_mock_response_with_default_variables(self, mock_codecov_client_class):
         mock_codecov_client_instance = Mock()
         mock_response = Mock()
@@ -166,7 +166,7 @@ class TestResultsEndpointTest(APITestCase):
             response_keys == serializer_fields
         ), f"Response keys {response_keys} don't match serializer fields {serializer_fields}"
 
-    @patch("sentry.codecov.endpoints.TestResults.test_results.CodecovApiClient")
+    @patch("sentry.codecov.endpoints.test_results.test_results.CodecovApiClient")
     def test_get_with_query_parameters(self, mock_codecov_client_class):
         mock_codecov_client_instance = Mock()
         mock_response = Mock()
@@ -211,7 +211,7 @@ class TestResultsEndpointTest(APITestCase):
         assert call_args[1]["variables"] == expected_variables
         assert response.status_code == 200
 
-    @patch("sentry.codecov.endpoints.TestResults.test_results.CodecovApiClient")
+    @patch("sentry.codecov.endpoints.test_results.test_results.CodecovApiClient")
     def test_get_with_term_filter(self, mock_codecov_client_class):
         mock_codecov_client_instance = Mock()
         mock_response = Mock()
@@ -255,7 +255,7 @@ class TestResultsEndpointTest(APITestCase):
         assert call_args[1]["variables"] == expected_variables
         assert response.status_code == 200
 
-    @patch("sentry.codecov.endpoints.TestResults.test_results.CodecovApiClient")
+    @patch("sentry.codecov.endpoints.test_results.test_results.CodecovApiClient")
     def test_get_with_cursor_alone_uses_default_limit_and_navigation(
         self, mock_codecov_client_class
     ):
@@ -294,7 +294,7 @@ class TestResultsEndpointTest(APITestCase):
         assert call_args[1]["variables"] == expected_variables
         assert response.status_code == 200
 
-    @patch("sentry.codecov.endpoints.TestResults.test_results.CodecovApiClient")
+    @patch("sentry.codecov.endpoints.test_results.test_results.CodecovApiClient")
     def test_get_with_cursor_and_direction(self, mock_codecov_client_class):
         mock_codecov_client_instance = Mock()
         mock_response = Mock()
